Update main-script.js
Still an issue with lag sometimes, not sure what causes it. If you need a scroll a ton disable the script and reload. At some point I will try to fix it. Appears to be related to how many CSS elements it affects slowing down the browser, I do not see a simple way to fix because of BMs log design.